refactor: zustand states
