+++ apps/web/src/components/Common/Layout.tsx
@@ -10,11 +10,11 @@ import { useTheme } from 'next-themes';
 import { type FC, type ReactNode } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { useFeatureFlagsStore } from 'src/store/non-persisted/useFeatureFlagsStore';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
+import { usePreferencesStore } from 'src/store/non-persisted/usePreferencesStore';
+import { useProStore } from 'src/store/non-persisted/useProStore';
 import { hydrateAuthTokens, signOut } from 'src/store/persisted/useAuthStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useNonceStore } from 'src/store/useNonceStore';
-import { usePreferencesStore } from 'src/store/usePreferencesStore';
-import { useProStore } from 'src/store/useProStore';
 import { useEffectOnce, useIsMounted } from 'usehooks-ts';
 import { isAddress } from 'viem';
 import { useAccount, useDisconnect } from 'wagmi';
