+++ apps/web/src/components/Common/Providers/LensSubscriptionsProvider.tsx
@@ -13,8 +13,8 @@ import getPushNotificationData from '@lib/getPushNotificationData';
 import { type FC } from 'react';
 import useWebSocket from 'react-use-websocket';
 import { isSupported, share } from 'shared-zustand';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
 import { signOut } from 'src/store/persisted/useAuthStore';
-import { useNonceStore } from 'src/store/useNonceStore';
 import { useNotificationPersistStore } from 'src/store/useNotificationPersistStore';
 import { useEffectOnce, useUpdateEffect } from 'usehooks-ts';
 import { isAddress } from 'viem';
