+++ apps/web/src/components/Composer/Actions/LivestreamSettings/index.tsx
@@ -2,7 +2,7 @@ import { VideoCameraIcon } from '@heroicons/react/24/outline';
 import { Tooltip } from '@hey/ui';
 import { motion } from 'framer-motion';
 import { type FC } from 'react';
-import { usePublicationStore } from 'src/store/usePublicationStore';
+import { usePublicationStore } from 'src/store/non-persisted/usePublicationStore';
 
 const LivestreamSettings: FC = () => {
   const showLiveVideoEditor = usePublicationStore(
