+++ apps/web/src/components/Composer/Actions/PollSettings/PollEditor.tsx
@@ -4,7 +4,7 @@ import { Button, Card, Input, Modal, Tooltip } from '@hey/ui';
 import plur from 'plur';
 import type { FC } from 'react';
 import { useState } from 'react';
-import { usePublicationStore } from 'src/store/usePublicationStore';
+import { usePublicationStore } from 'src/store/non-persisted/usePublicationStore';
 
 const PollEditor: FC = () => {
   const setShowPollEditor = usePublicationStore(
