+++ apps/web/src/components/Profile/Feed.tsx
@@ -14,7 +14,7 @@ import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
 import { ProfileFeedType } from 'src/enums';
 import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
-import { useProfileFeedStore } from 'src/store/useProfileFeedStore';
+import { useProfileFeedStore } from 'src/store/non-persisted/useProfileFeedStore';
 
 interface FeedProps {
   profile: Profile;
