+++ apps/web/src/components/Profile/Filters/MediaFilter.tsx
@@ -4,7 +4,7 @@ import { AdjustmentsVerticalIcon } from '@heroicons/react/24/outline';
 import { Checkbox, Tooltip } from '@hey/ui';
 import cn from '@hey/ui/cn';
 import type { ChangeEvent } from 'react';
-import { useProfileFeedStore } from 'src/store/useProfileFeedStore';
+import { useProfileFeedStore } from 'src/store/non-persisted/useProfileFeedStore';
 
 const MediaFilter = () => {
   const mediaFeedFilters = useProfileFeedStore(
