+++ apps/web/src/components/Profile/NftGallery/Gallery.tsx
@@ -20,12 +20,12 @@ import cn from '@hey/ui/cn';
 import type { FC } from 'react';
 import { useState } from 'react';
 import { toast } from 'react-hot-toast';
-import useProfileStore from 'src/store/persisted/useProfileStore';
-import type { NftGalleryItem } from 'src/store/useNftGalleryStore';
+import type { NftGalleryItem } from 'src/store/non-persisted/useNftGalleryStore';
 import {
   GALLERY_DEFAULTS,
   useNftGalleryStore
-} from 'src/store/useNftGalleryStore';
+} from 'src/store/non-persisted/useNftGalleryStore';
+import useProfileStore from 'src/store/persisted/useProfileStore';
 
 import Create from './Create';
 import NftCard from './NftCard';
