+++ apps/web/src/components/Publication/Actions/Share/Mirror.tsx
@@ -28,8 +28,8 @@ import { Leafwatch } from '@lib/leafwatch';
 import { type FC, useState } from 'react';
 import { toast } from 'react-hot-toast';
 import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useNonceStore } from 'src/store/useNonceStore';
 import { useContractWrite, useSignTypedData } from 'wagmi';
 
 interface MirrorProps {
