+++ apps/web/src/components/Publication/LensOpenActions/CollectModule/CollectAction.tsx
@@ -37,8 +37,8 @@ import type { FC } from 'react';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
 import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useNonceStore } from 'src/store/useNonceStore';
 import { isAddress } from 'viem';
 import {
   useAccount,
