+++ apps/web/src/components/Publication/PublicationHeader.tsx
@@ -7,7 +7,7 @@ import stopEventPropagation from '@hey/lib/stopEventPropagation';
 import cn from '@hey/ui/cn';
 import { type FC } from 'react';
 import { useFeatureFlagsStore } from 'src/store/non-persisted/useFeatureFlagsStore';
-import { usePublicationStore } from 'src/store/usePublicationStore';
+import { usePublicationStore } from 'src/store/non-persisted/usePublicationStore';
 
 import PublicationMenu from './Actions/Menu';
 import Source from './Source';
