+++ apps/web/src/components/Settings/Manager/ProfileManager/Managers/List.tsx
@@ -21,8 +21,8 @@ import { type FC, useState } from 'react';
 import { useInView } from 'react-cool-inview';
 import toast from 'react-hot-toast';
 import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useNonceStore } from 'src/store/useNonceStore';
 import type { Address } from 'viem';
 import { useContractWrite, useSignTypedData } from 'wagmi';
