+++ apps/web/src/components/Settings/Preferences/IsPride.tsx
@@ -6,7 +6,7 @@ import { Leafwatch } from '@lib/leafwatch';
 import axios from 'axios';
 import type { FC } from 'react';
 import { toast } from 'react-hot-toast';
-import { usePreferencesStore } from 'src/store/usePreferencesStore';
+import { usePreferencesStore } from 'src/store/non-persisted/usePreferencesStore';
 
 const IsPride: FC = () => {
   const isPride = usePreferencesStore((state) => state.isPride);
