+++ apps/web/src/components/Shared/Audio/index.tsx
@@ -7,7 +7,7 @@ import { Leafwatch } from '@lib/leafwatch';
 import type { APITypes } from 'plyr-react';
 import type { ChangeEvent, FC } from 'react';
 import { useRef, useState } from 'react';
-import { usePublicationStore } from 'src/store/usePublicationStore';
+import { usePublicationStore } from 'src/store/non-persisted/usePublicationStore';
 import { object, string } from 'zod';
 
 import CoverImage from './CoverImage';
