+++ apps/web/src/hooks/useCreatePublication.tsx
@@ -32,10 +32,10 @@ import checkDispatcherPermissions from '@hey/lib/checkDispatcherPermissions';
 import getSignature from '@hey/lib/getSignature';
 import { OptmisticPublicationType } from '@hey/types/enums';
 import { useRouter } from 'next/router';
+import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
+import { usePublicationStore } from 'src/store/non-persisted/usePublicationStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
 import { useTransactionStore } from 'src/store/persisted/useTransactionStore';
-import { useNonceStore } from 'src/store/useNonceStore';
-import { usePublicationStore } from 'src/store/usePublicationStore';
 import { useContractWrite, useSignTypedData } from 'wagmi';
 
 interface CreatePublicationProps {
