+++ apps/web/src/store/non-persisted/useNftGalleryStore.ts
